[Embeddable] remove Container class (#204105)
PR does the following
* Removes Container class and IContainer interface
* Removes parent from Embeddable class
* Removes more dead code from embeddable plugin like EmbeddablePanel
component
* Cleans up all usages of Container and IContainer

---------

(cherry picked from commit 08e7298cc0cb086081361d471d4226db0856d833)
