+++ src/plugins/dashboard/public/dashboard_app/top_nav/dashboard_editing_toolbar.tsx
@@ -80,12 +80,6 @@ export function DashboardEditingToolbar({ isDisabled }: { isDisabled?: boolean }
     [dashboardApi]
   );
 
-  /**
-   * embeddableFactory: Required, you can get the factory from embeddableStart.getEmbeddableFactory(<embeddable type, i.e. lens>)
-   * initialInput: Optional, use it in case you want to pass your own input to the factory
-   * dismissNotification: Optional, if not passed a toast will appear in the dashboard
-   */
-
   const controlGroupApi = useStateFromPublishingSubject(dashboardApi.controlGroupApi$);
   const extraButtons = [
     <EditorMenu createNewVisType={createNewVisType} isDisabled={isDisabled} />,
