+++ src/plugins/embeddable/public/index.ts
@@ -12,11 +12,9 @@ import { EmbeddablePublicPlugin } from './plugin';
 
 export { useAddFromLibraryTypes } from './add_from_library/registry';
 export { openAddFromLibraryFlyout } from './add_from_library/open_add_from_library_flyout';
-export { EmbeddablePanel } from './embeddable_panel';
 export {
   cellValueTrigger,
   CELL_VALUE_TRIGGER,
-  Container,
   contextMenuTrigger,
   CONTEXT_MENU_TRIGGER,
   defaultEmbeddableFactoryProvider,
@@ -58,9 +56,6 @@ export type {
   Adapters,
   CellValueContext,
   ChartActionContext,
-  ContainerInput,
-  ContainerOutput,
-  EmbeddableContainerSettings,
   EmbeddableContext,
   EmbeddableEditorState,
   EmbeddableFactory,
@@ -70,11 +65,9 @@ export type {
   EmbeddableOutput,
   EmbeddablePackageState,
   FilterableEmbeddable,
-  IContainer,
   IEmbeddable,
   MultiValueClickContext,
   OutputSpec,
-  PanelState,
   PropertySpec,
   RangeSelectContext,
   ReferenceOrValueEmbeddable,
