--- src/plugins/embeddable/public/lib/containers/index.ts
@@ -1,17 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-export type {
-  IContainer,
-  PanelState,
-  ContainerInput,
-  ContainerOutput,
-  EmbeddableContainerSettings,
-} from './i_container';
-export { Container } from './container';
