+++ src/plugins/embeddable/public/lib/index.ts
@@ -11,7 +11,6 @@ export * from './errors';
 export * from './embeddables';
 export * from './types';
 export * from './triggers';
-export * from './containers';
 export * from './state_transfer';
 export * from './reference_or_value_embeddable';
 export * from './self_styled_embeddable';
