--- src/plugins/embeddable/public/tests/fixtures/hello_world_embeddable_react.tsx
@@ -1,17 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-import React from 'react';
-import { HelloWorldEmbeddable } from './hello_world_embeddable';
-
-export class HelloWorldEmbeddableReact extends HelloWorldEmbeddable {
-  public render() {
-    return <div data-test-subj="helloWorldEmbeddable">HELLO WORLD!</div>;
-  }
-}
