+++ src/plugins/visualizations/public/legacy/embeddable/visualize_embeddable.tsx
@@ -27,7 +27,6 @@ import {
   EmbeddableInput,
   EmbeddableOutput,
   FilterableEmbeddable,
-  IContainer,
   ReferenceOrValueEmbeddable,
   SavedObjectEmbeddableInput,
 } from '@kbn/embeddable-plugin/public';
@@ -41,18 +40,28 @@ import type { RenderMode } from '@kbn/expressions-plugin/common';
 import { DATA_VIEW_SAVED_OBJECT_TYPE } from '@kbn/data-views-plugin/public';
 import { mapAndFlattenFilters } from '@kbn/data-plugin/public';
 import { isChartSizeEvent } from '@kbn/chart-expressions-common';
+import { StartServicesGetter } from '@kbn/kibana-utils-plugin/public';
 import { isFallbackDataView } from '../../visualize_app/utils';
 import { VisualizationMissedSavedObjectError } from '../../components/visualization_missed_saved_object_error';
 import VisualizationError from '../../components/visualization_error';
 import { VISUALIZE_EMBEDDABLE_TYPE } from './constants';
 import { SerializedVis, Vis } from '../../vis';
-import { getApplication, getExecutionContext, getExpressions, getUiActions } from '../../services';
+import { getApplication, getExpressions, getUiActions } from '../../services';
 import { VIS_EVENT_TO_TRIGGER } from '../../embeddable/events';
-import { VisualizeEmbeddableFactoryDeps } from './visualize_embeddable_factory';
 import { getSavedVisualization } from '../../utils/saved_visualize_utils';
 import { VisSavedObject } from '../../types';
 import { toExpressionAst } from '../../embeddable/to_ast';
 import { AttributeService } from './attribute_service';
+import { VisualizationsStartDeps } from '../../plugin';
+
+export interface VisualizeEmbeddableDeps {
+  start: StartServicesGetter<
+    Pick<
+      VisualizationsStartDeps,
+      'inspector' | 'embeddable' | 'data' | 'savedObjectsTaggingOss' | 'spaces'
+    >
+  >;
+}
 
 export interface VisualizeEmbeddableConfiguration {
   vis: Vis;
@@ -60,7 +69,7 @@ export interface VisualizeEmbeddableConfiguration {
   editPath: string;
   editUrl: string;
   capabilities: { visualizeSave: boolean; dashboardSave: boolean; visualizeOpen: boolean };
-  deps: VisualizeEmbeddableFactoryDeps;
+  deps: VisualizeEmbeddableDeps;
 }
 
 export interface VisualizeInput extends EmbeddableInput {
@@ -120,7 +129,7 @@ export class VisualizeEmbeddable
   private warningDomNode: any;
   public readonly type = VISUALIZE_EMBEDDABLE_TYPE;
   private abortController?: AbortController;
-  private readonly deps: VisualizeEmbeddableFactoryDeps;
+  private readonly deps: VisualizeEmbeddableDeps;
   private readonly inspectorAdapters?: Adapters;
   private attributeService?: AttributeService<
     VisualizeSavedObjectAttributes,
@@ -140,22 +149,17 @@ export class VisualizeEmbeddable
       VisualizeSavedObjectAttributes,
       VisualizeByValueInput,
       VisualizeByReferenceInput
-    >,
-    parent?: IContainer
+    >
   ) {
-    super(
-      initialInput,
-      {
-        defaultTitle: vis.title,
-        defaultDescription: vis.description,
-        editPath,
-        editApp: 'visualize',
-        editUrl,
-        indexPatterns,
-        visTypeName: vis.type.name,
-      },
-      parent
-    );
+    super(initialInput, {
+      defaultTitle: vis.title,
+      defaultDescription: vis.description,
+      editPath,
+      editApp: 'visualize',
+      editUrl,
+      indexPatterns,
+      visTypeName: vis.type.name,
+    });
     this.deps = deps;
     this.timefilter = timefilter;
     this.syncColors = this.input.syncColors;
@@ -270,8 +274,6 @@ export class VisualizeEmbeddable
 
         this.vis.uiState.on('change', this.uiStateChangeHandler);
       }
-    } else if (this.parent) {
-      this.vis.uiState.clearAllKeys();
     }
   }
 
@@ -572,7 +574,6 @@ export class VisualizeEmbeddable
   };
 
   private getExecutionContext() {
-    const parentContext = this.parent?.getInput().executionContext || getExecutionContext().get();
     const child: KibanaExecutionContext = {
       type: 'agg_based',
       name: this.vis.type.name,
@@ -582,7 +583,6 @@ export class VisualizeEmbeddable
     };
 
     return {
-      ...parentContext,
       child,
     };
   }
