+++ x-pack/plugins/discover_enhanced/public/actions/explore_data/explore_data_chart_action.test.ts
@@ -7,15 +7,10 @@
 import { coreMock } from '@kbn/core/public/mocks';
 import { DataView } from '@kbn/data-views-plugin/common';
 import { DiscoverAppLocator } from '@kbn/discover-plugin/common';
-import { ViewMode } from '@kbn/embeddable-plugin/public';
 import type { Filter, RangeFilter } from '@kbn/es-query';
 import { i18n } from '@kbn/i18n';
-import { ViewMode as ViewModeType } from '@kbn/presentation-publishing';
+import { ViewMode } from '@kbn/presentation-publishing';
 import { sharePluginMock } from '@kbn/share-plugin/public/mocks';
-import {
-  VisualizeEmbeddableContract,
-  VISUALIZE_EMBEDDABLE_TYPE,
-} from '@kbn/visualizations-plugin/public';
 import { BehaviorSubject } from 'rxjs';
 import { Params, PluginDeps } from './abstract_explore_data_action';
 import { ExploreDataChartAction, ExploreDataChartActionContext } from './explore_data_chart_action';
@@ -70,18 +65,18 @@ const setup = (
   };
   const action = new ExploreDataChartAction(params);
 
-  const embeddable: VisualizeEmbeddableContract = {
-    type: VISUALIZE_EMBEDDABLE_TYPE,
-    dataViews: new BehaviorSubject([
+  const embeddable = {
+    type: 'anyEmbeddable',
+    dataViews: new BehaviorSubject<undefined | DataView[]>([
       {
         id: 'index-ptr-foo',
-      },
+      } as DataView,
     ]),
-    filters$: new BehaviorSubject([]),
+    filters$: new BehaviorSubject<Filter[]>([]),
     parentApi: {
-      viewMode: new BehaviorSubject(ViewMode.VIEW),
+      viewMode: new BehaviorSubject<ViewMode>('view'),
     },
-  } as unknown as VisualizeEmbeddableContract;
+  };
 
   const context = {
     filters,
@@ -138,11 +133,11 @@ describe('"Explore underlying data" panel action', () => {
       embeddable.dataViews = new BehaviorSubject<undefined | DataView[]>([
         {
           id: 'index-ptr-foo',
-        },
+        } as DataView,
         {
           id: 'index-ptr-bar',
-        },
-      ] as any as DataView[]);
+        } as DataView,
+      ]);
 
       const isCompatible = await action.isCompatible(context);
 
@@ -171,7 +166,7 @@ describe('"Explore underlying data" panel action', () => {
     test('returns false if dashboard is in edit mode', async () => {
       const { action, embeddable, context } = setup();
       if (embeddable.parentApi) {
-        embeddable.parentApi.viewMode = new BehaviorSubject<ViewModeType>(ViewMode.EDIT);
+        embeddable.parentApi.viewMode = new BehaviorSubject<ViewMode>('edit');
       }
       const isCompatible = await action.isCompatible(context);
